Code cleanup and optimize imports.

+++ wf-server/src/main/java/ru/taskurotta/server/config/expiration/impl/TimeoutWithFixedRetryPolicy.java
@@ -1,12 +1,12 @@
 package ru.taskurotta.server.config.expiration.impl;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import java.util.Map;
 import java.util.Properties;
 import java.util.UUID;
 import java.util.concurrent.ConcurrentHashMap;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 
 public class TimeoutWithFixedRetryPolicy extends TimeoutPolicy {
@@ -35,7 +35,7 @@ public class TimeoutWithFixedRetryPolicy extends TimeoutPolicy {
             Integer taskRetry = expirations.get(uuid);
 
             if(taskRetry == null || taskRetry < retry) {
-                expirations.put(uuid, Integer.valueOf(taskRetry==null?1: taskRetry.intValue()+1));
+                expirations.put(uuid, taskRetry == null ? 1 : taskRetry + 1);
             } else {
                 result = false;
                 logger.error("Task[{}] expiration policy commit failed: Task has been already retried for [{}]/[{}] times", uuid, taskRetry, retry);
